Support for running only 1 test

+++ Testing/copier.sh
@@ -2,10 +2,12 @@
 
 FILES=(`ls -I copier.sh -I cleaner.sh`)
 
-for i in ${FILES[*]}; do
-	for j in $(seq 1 $1); do
-		NAME="${i::-4}"
-		cp $i ${NAME}_${j}.txt
+if [[ $1 -ge 1 ]]; then
+	for i in ${FILES[*]}; do
+		for j in $(seq 1 $1); do
+			NAME="${i::-4}"
+			cp $i ${NAME}_${j}.txt
+		done
+		rm $i
 	done
-	rm $i
-done
+fi
